requirements.txt
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.2.0
astatine==0.3.3
astor==0.8.1
astpretty==3.0.0
astroid==2.15.8
asttokens==2.4.1
attrs==23.2.0
autoflake==1.7.8
Babel==2.14.0
bandit==1.7.7
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cognitive-complexity==1.3.0
colorama==0.4.6
coverage==7.4.1
darglint==1.8.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
dlint==0.14.1
dnspython==2.4.2
doc8==1.1.1
docformatter==1.7.5
docutils==0.19
domdf-python-tools==3.8.0.post2
ecdsa==0.18.0
email-validator==2.1.0.post1
eradicate==2.3.0
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
flake8==5.0.4
flake8-2020==1.8.1
flake8-aaa==0.17.0
flake8-annotations==3.0.1
flake8-annotations-complexity==0.0.8
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-encodings==0.5.1
flake8-eradicate==1.5.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.8
flake8-functions-names==0.4.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.2
flake8-isort==6.1.1
flake8-literal==1.4.0
flake8-logging-format==0.9.0
flake8-markdown==0.5.0
flake8-mutable==1.2.0
flake8-no-pep420==2.7.0
flake8-noqa==1.4.0
flake8-pie==0.16.0
flake8-plugin-utils==1.3.3
flake8-pyi==22.11.0
flake8-pylint==0.2.1
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.1
flake8-string-format==0.3.0
flake8-tidy-imports==4.10.0
flake8-typing-imports==1.15.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.6
flake8-warnings==0.4.0
flake8_simplify==0.21.0
h11==0.14.0
hypothesis==6.97.0
hypothesmith==0.1.9
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-flake8==0.7.1
jupyterlab_pygments==0.3.0
lark-parser==0.12.0
lazy-object-proxy==1.10.0
libcst==0.4.10
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mr-proper==0.0.7
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
packaging==23.2
pandas-vet==0.2.3
pandocfilters==1.5.1
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.13.3
platformdirs==4.1.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pybetter==0.4.1
pycln==2.4.0
pycodestyle==2.9.1
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
Pygments==2.17.2
pylint==2.17.7
pyproject-api==1.6.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-dev-tools==2023.3.24
python-dotenv==1.0.0
python-jose==3.3.0
pyupgrade==3.15.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
removestar==1.5
requests==2.31.0
restructuredtext-lint==1.4.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.25
ssort==0.12.3
starlette==0.35.1
stdlib-list==0.10.0
stevedore==5.1.0
termcolor==2.4.0
tinycss2==1.2.1
tokenize-rt==5.2.0
tomlkit==0.12.3
tornado==6.4
tox==4.12.1
tox-travis==0.12
traitlets==5.14.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
untokenize==0.1.1
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.0
webencodings==0.5.1
wrapt==1.16.0
